MB-27661 [10/n]: Fix MSVC warnings - C4267
This patch addresses the following generated warnings:

C4267 - var : conversion from 'size_t' to 'type', possible loss of data
        The compiler detected a conversion from size_t to a smaller type.

Since there are a lot of these warnings, use multiple commits to sort
them out to make reviewing easier.

Reviewed-on: http://review.couchbase.org/90539
Tested-by: Build Bot <build@couchbase.com>

+++ cbsasl/client_server_test.cc
@@ -22,6 +22,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <array>
+#include <gsl/gsl>
 
 const char* cbpwfile = "cbsasl_test.pw";
 
@@ -75,7 +76,7 @@ static int sasl_get_cnonce(void* context,
 
     struct my_sasl_ctx* ctx = reinterpret_cast<my_sasl_ctx*>(context);
     *result = ctx->nonce.data();
-    *len = ctx->nonce.length();
+    *len = gsl::narrow<unsigned int>(ctx->nonce.length());
 
     return CBSASL_OK;
 }
